requirements.txt
amqp==5.0.2
appdirs==1.4.4
asgiref==3.2.10
billiard==3.6.3.0
celery==5.0.2
certifi==2019.9.11
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
conda==4.3.16
coverage==5.3
distlib==0.3.1
Django==3.1
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-extensions==3.0.5
djangorestframework==3.11.1
djangorestframework-simplejwt==4.6.0
Faker==4.1.2
filelock==3.0.12
flake8==3.8.3
flake8-builtins==1.5.3
flake8-import-order==0.18.1
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
install==1.3.3
kombu==5.0.2
mccabe==0.6.1
Pillow==7.2.0
pipenv==2018.11.26
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycosat==0.6.3
pyflakes==2.2.0
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.10
vine==5.0.0
virtualenv==16.7.7
virtualenv-clone==0.5.3
wcwidth==0.2.5
zipp==3.1.0
